challenge-1/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "warehouse-picker-vla"
version = "0.1.0"
description = "Agentic VLA warehouse picking robot system"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "torchvision",
    "transformers",
    "accelerate",
    "pillow",
    "pyyaml",
    "pyzmq",
    "rich",
    "matplotlib",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-timeout"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring MPS/GPU",
    "docker: marks tests requiring Docker",
    "vision: marks tests requiring claude -p vision",
]
